[HRM-3] implement tests and some code refactoring
type: tests & code

description: -

+++ auth-service/src/main/java/org/example/api/constant/JwtProperties.java
@@ -12,6 +12,6 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 public class JwtProperties {
 
     private final String secretKey;
-    private final long validityInMinutes;
+    private final int validityInMinutes;
 
 }
